+++ tests/unittests/test_error_handling.py
@@ -1,7 +1,8 @@
 import unittest
+import requests
+
 from unittest import mock
 from parameterized import parameterized
-import requests
 
 from tap_mixpanel import client
 
@@ -130,7 +131,9 @@ class TestMixpanelErrorHandling(unittest.TestCase):
         with self.assertRaises(error) as e:
             mock_client.perform_request("GET")
 
-        expected_error_message = f"HTTP-error-code: {error_code}, Error: {error_message}"
+        expected_error_message = (
+            f"HTTP-error-code: {error_code}, Error: {error_message}"
+        )
 
         # Verifying the message formed for the custom exception
         self.assertEqual(str(e.exception), expected_error_message)
@@ -140,6 +143,7 @@ class TestMixpanelErrorHandling(unittest.TestCase):
         ["400 different timezone error", 400, mock_400_different_timezone(), client.MixpanelBadRequestError, "A validation exception has occurred. Please validate the timezone with the MixPanel UI under project settings."],
         ["400 timeout error", 400, MockResponse(400, text=timeout_400_error), client.MixpanelBadRequestError, "Timeout Error.(Please verify your credentials.)"],
         ["401 error", 401, MockResponse(401), client.MixpanelUnauthorizedError, "Invalid authorization credentials."],
+       # ["403 error", 403, MockResponse(403), client.MixpanelForbiddenError, "User is not a member of this project: project_id or this project is invalid"],
         ["404 error", 404, MockResponse(404), client.MixpanelNotFoundError, "The resource you have specified cannot be found."],
         ["404 error", 404, mock_send_error(), client.MixpanelNotFoundError, "Resource not found error message from API response field 'error'."],
         ["404 error", 404, mock_send_message(), client.MixpanelNotFoundError, "Resource not found error message from API response field 'message'."],
@@ -166,15 +170,19 @@ class TestMixpanelErrorHandling(unittest.TestCase):
         with self.assertRaises(error) as e:
             mock_client.check_access()
 
-        expected_error_message = f"HTTP-error-code: {error_code}, Error: {error_message}"
+        expected_error_message = (
+            f"HTTP-error-code: {error_code}, Error: {error_message}"
+        )
 
         # Verifying the message formed for the custom exception
         self.assertEqual(str(e.exception), expected_error_message)
 
-    @parameterized.expand([
-        ["500 error", MockResponse(500), client.MixpanelInternalServiceError],
-        ["501 error", MockResponse(501), client.Server5xxError],
-    ])
+    @parameterized.expand(
+        [
+            ["500 error", MockResponse(500), client.MixpanelInternalServiceError],
+            ["501 error", MockResponse(501), client.Server5xxError],
+        ]
+    )
     @mock.patch("requests.Session.request")
     def test_request_with_handling_for_5xx_exception_handling(
         self, test_name, mock_response, error, mock_request, mock_sleep
@@ -209,6 +217,54 @@ class TestMixpanelErrorHandling(unittest.TestCase):
             request_timeout=REQUEST_TIMEOUT,
         )
         with self.assertRaises(client.ReadTimeoutError):
+            mock_client.check_access()
+
+        # Verify that requests.Session.request is called 5 times
+        self.assertEqual(mock_request.call_count, 5)
+
+    @mock.patch("requests.Session.request")
+    @mock.patch("tap_mixpanel.client.LOGGER.warning")
+    def test_check_access_402_exception_handling(
+        self, mock_logger, mock_request, mock_sleep
+    ):
+        """
+        Test that `check_access` method does not throw 402 error and prints
+        warning logger instead.
+        """
+        mock_request.return_value = MockResponse(402)
+        mock_client = client.MixpanelClient(
+            api_secret="mock_api_secret",
+            service_account_username="mock_service_account_username",
+            service_account_secret="service_account_secret",
+            project_id="project_id",
+            api_domain="mock_api_domain",
+            request_timeout=REQUEST_TIMEOUT,
+        )
+
+        mock_client.check_access()
+
+        # Verify that for 402 error expected logger is printed.
+        mock_logger.assert_called_with(
+            "Mixpanel returned a 402 from the Engage API. Engage stream will be skipped."
+        )
+
+
+# Mock time.sleep to reduce the time
+@mock.patch("time.sleep", return_value=None)
+class TestMixpanelConnectionResetErrorHandling(unittest.TestCase):
+
+    @mock.patch("requests.Session.request", side_effect=requests.models.ProtocolError)
+    def test_check_access_handle_timeout_error(self, mock_request, mock_time):
+        """
+        Check whether the request backoffs properly for `check_access` method for 5 times in case of Timeout error.
+        """
+        mock_client = client.MixpanelClient(api_secret="mock_api_secret",
+                                            service_account_username="mock_service_account_username",
+                                            service_account_secret="service_account_secret",
+                                            project_id="project_id",
+                                            api_domain="mock_api_domain",
+                                            request_timeout=REQUEST_TIMEOUT)
+        with self.assertRaises(requests.models.ProtocolError):
             mock_client.check_access()
 
         # Verify that requests.Session.request is called 5 times